+++ scim2_models/rfc7644/patch_op.py
@@ -31,6 +31,6 @@ class PatchOperation(ComplexAttribute):
 class PatchOp(Message):
     schemas: List[str] = ["urn:ietf:params:scim:api:messages:2.0:PatchOp"]
 
-    operations: List[PatchOperation] = Field(..., alias="Operations")
+    operations: List[PatchOperation] = Field(None, alias="Operations")
     """The body of an HTTP PATCH request MUST contain the attribute
     "Operations", whose value is an array of one or more PATCH operations."""